Seperated blocking and queues task into 2 projects for ease of life
