+++ projects/task1/src/main.c
@@ -3,80 +3,48 @@
 
 #include "FreeRTOS.h"
 #include "tasks.h"
-#include "queues.h"
-#include "status.h"
-#include "delay.h"
 
+#include "gpio.h"
 #include "log.h"
 #include "misc.h"
+#include "delay.h"
 
-
-#define ITEM_SZ 6
-#define QUEUE_LEN 5
-#define BUF_SIZE (QUEUE_LEN * ITEM_SZ)
-
-static const char s_list[QUEUE_LEN][ITEM_SZ] = {
-	"Item1",
-	"Item2",
-	"Item3",
-	"Item4",
-	"Item5"
-};
-
-// Task static entities
-static uint8_t s_queue1_buf[BUF_SIZE];
-static Queue s_queue1 = {
-  // Add parameters
-  .num_items = QUEUE_LEN,
-  .item_size = ITEM_SZ,
-  .storage_buf = s_queue1_buf,
-};
-
+// Non blocking delay. Simply consumes cpu cycles until a given time has passed
+// static void prv_delay(const TickType_t delay_ms) {
+  // TickType_t curr_tick = xTaskGetTickCount();
+  // while(xTaskGetTickCount() - curr_tick < pdMS_TO_TICKS(delay_ms))
+ // {}
+// }
 
 TASK(task1, TASK_STACK_512) {
-  LOG_DEBUG("Task 1 initialized!\n");
-  StatusCode ret;
+  int counter1 = 0;
   while (true) {
-    // Your code goes here
-    for (int i = 0; i < QUEUE_LEN; i++){
-      ret = queue_send(&s_queue1, s_list[i], 0);
-      if (ret != STATUS_CODE_OK){
-        LOG_DEBUG("Write to queue failed\n");
-      }
-      delay_ms(1000);
-    }
+  // Your code here
+  LOG_DEBUG("This is task 1 and the counter value is: %d \n", counter1);
+  counter1++;
+  delay_ms(1000); //Delay function from delay.h instead of calling prv_delay (I hope this is what was expected from of the task)
   }
 }
 
 TASK(task2, TASK_STACK_512) {
-  LOG_DEBUG("Task 2 initialized!\n");
-  const char outstr[ITEM_SZ];
-  StatusCode ret;
+  int counter2 = 0;
   while (true) {
-    // Your code goes here
-     for (int i = 0; i < QUEUE_LEN; i++){
-      ret = queue_receive(&s_queue1, outstr, 0);
-      if (ret != STATUS_CODE_OK){
-        LOG_DEBUG("Read from queue failed\n");
-      }else{
-        LOG_DEBUG("String Read from queue: %s\n", outstr);
-      }
-      delay_ms(1000);
-     }
+  // Your code here
+  LOG_DEBUG("This is task 2 and the counter value is: %d \n", counter2);
+  counter2++;
+  delay_ms(1000); //Delay function from delay.h instead of calling prv_delay (I hope this is what was expected from of the task)
   }
 }
 
 int main(void) {
-  log_init();
-  tasks_init();
-  // Initialize queues here
-  queue_init(&s_queue1);
+    log_init();
+    tasks_init();
+    // Create tasks here
+    tasks_init_task(task1, TASK_PRIORITY(2), NULL);
+    tasks_init_task(task2, TASK_PRIORITY(1), NULL);
 
-  tasks_init_task(task1, TASK_PRIORITY(2), NULL);
-  tasks_init_task(task2, TASK_PRIORITY(2), NULL);
-
-  LOG_DEBUG("Program start...\n");
-  tasks_start();
-
-  return 0;
+    LOG_DEBUG("Program start...\n");
+    // Start the scheduler
+    tasks_start();
+    return 0;
 }
