+++ projects/task2/config.json
@@ -1,0 +1,6 @@
+{
+    "libs": [
+        "FreeRTOS",
+        "ms-common"
+    ]
+}
